+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Add the `react4j-drumloop` sample application into set off used to track code size changes over time.
 * Upgrade the `org.realityforge.com.google.elemental2:*` libraries to version `1.0.0-b16-6897368`.
+* **\[core\]** Allow the invocation of `DisposeNotifier.removeOnDisposeListener(...)` when the
+  `DisposeNotifier` has been disposed to avoid per-component generation to check whether `DisposeNotifier`
+  has been disposed in the rare circumstances where it is required.
 
 ### [v0.129](https://github.com/arez/arez/tree/v0.129) (2019-02-11)
 [Full Changelog](https://github.com/arez/arez/compare/v0.128...v0.129)